Split up filteredData getters tests file, and added missing tests
